# port masked_fill from TH to ATen