--- a/aten/src/ATen/Declarations.cwrap
+++ b/aten/src/ATen/Declarations.cwrap
@@ -20,12 +20,12 @@
 ]]
 [[
   name: _th_masked_fill_
-  cpu_bool: True
   cuda_bool: True
-  cpu_bfloat16: True
   cuda_bfloat16: True
   cname: maskedFill
   variants: function
+  backends:
+    - CUDA
   return: self
   options:
     - arguments:
@@ -35,12 +35,12 @@
 ]]
 [[
   name: _th_masked_fill_bool_
-  cpu_bool: True
   cuda_bool: True
-  cpu_bfloat16: True
   cuda_bfloat16: True
   cname: maskedFillBool
   variants: function
+  backends:
+    - CUDA
   return: self
   options:
     - arguments: