--- a/aten/src/ATen/native/LegacyDefinitions.cpp
+++ b/aten/src/ATen/native/LegacyDefinitions.cpp
@@ -8,43 +8,6 @@ namespace at { namespace native {
 
 // Methods
 
-Tensor & masked_fill__cpu(Tensor& self, const Tensor & mask, Scalar value) {
-  auto maybe_outnames = namedinference::broadcast_to_outnames(self, mask, "masked_fill_");
-  Tensor b_mask;
-  std::tie(b_mask) = expand_inplace(self, mask, "masked_fill_");
-  // As we dispatch on self and TH is type-checked, we need different definitions.
-  // This can be fixed by moving to ATen.
-  if (b_mask.dtype() == at::ScalarType::Byte) {
-    AT_WARN("masked_fill_ received a mask with dtype torch.uint8, this behavior is now deprecated," \
-            "please use a mask with dtype torch.bool instead.");
-    legacy::cpu::_th_masked_fill_(self, b_mask, value);
-  } else {
-    legacy::cpu::_th_masked_fill_bool_(self, b_mask, value);
-  }
-  namedinference::propagate_names_if_nonempty(self, maybe_outnames);
-  return self;
-}
-
-Tensor & masked_fill__cpu(Tensor& self, const Tensor & mask, const Tensor & value) {
-  auto maybe_outnames = namedinference::broadcast_to_outnames(self, mask, "masked_fill_");
-
-  TORCH_CHECK(value.dim() == 0, "masked_fill_ only supports a 0-dimensional value tensor, but got tensor "
-      "with ", value.dim(), " dimension(s).");
-  Tensor b_mask;
-  std::tie(b_mask) = expand_inplace(self, mask, "masked_fill_");
-  // As we dispatch on self and TH is type-checked, we need different definitions.
-  // This can be fixed by moving to ATen.
-  if (b_mask.dtype() == at::ScalarType::Byte) {
-    AT_WARN("masked_fill_ received a mask with dtype torch.uint8, this behavior is now deprecated," \
-            "please use a mask with dtype torch.bool instead.");
-    legacy::cpu::_th_masked_fill_(self, b_mask, value.item());
-  } else {
-    legacy::cpu::_th_masked_fill_bool_(self, b_mask, value.item());
-  }
-  namedinference::propagate_names_if_nonempty(self, maybe_outnames);
-  return self;
-}
-
 Tensor & masked_scatter__cpu(Tensor& self, const Tensor & mask, const Tensor & source) {
   Tensor b_mask;
   std::tie(b_mask) = expand_inplace(self, mask, "masked_scatter_");