port `masked_fill` from TH to ATen with TensorIterator.

single core performance roughly stays the same, single socket performance has **3~16x** boost.

`masked_fill` is missing from https://github.com/pytorch/pytorch/issues/24507