@mingfeima 2020-02-14
Below shows the performance compare between **original** and **this pr**, benchmark code available at this [op_bench-py](https://github.com/mingfeima/op_bench-py), to reproduce, `./run.sh masked_fillpy`. Test machine is Xeon Skylake 6148 with 2*20 cores @2.40GHz.

The unit is `ms per iteration`, the lower the better.
#### single socket run

input size | contiguous | original | this pr | speedup
-- | -- | -- | -- | --
[128 1000] | TRUE | 0.481 | 0.134 | 3.59
[256 1000] | TRUE | 0.979 | 0.134 | 7.31
[512 1000] | TRUE | 1.021 | 0.149 | 6.85
[1024 1000] | TRUE | 1.046 | 0.239 | 4.38
[128 1000] | FALSE | 0.507 | 0.134 | 3.78
[256 1000] | FALSE | 1.017 | 0.135 | 7.53
[512 1000] | FALSE | 2.029 | 0.150 | 13.53
[1024 1000] | FALSE | 4.057 | 0.241 | 16.83


#### single core run

input size | contiguous | original | this pr | speedup
-- | -- | -- | -- | --
[128 1000] | TRUE | 0.481 | 0.493 | 0.98
[256 1000] | TRUE | 0.999 | 0.994 | 1.01
[512 1000] | TRUE | 1.925 | 1.980 | 0.97
[1024 1000] | TRUE | 3.859 | 3.953 | 0.98
[128 1000] | FALSE | 0.514 | 0.517 | 0.99
[256 1000] | FALSE | 1.028 | 0.997 | 1.03
[512 1000] | FALSE | 2.029 | 2.000 | 1.01
[1024 1000] | FALSE | 4.157 | 4.107 | 1.01